+++ Demo/DemoMultiTarget/ViewController.swift
@@ -56,14 +56,14 @@ class ViewController: UITableViewController {
     func uploadGiphy() {
         let data = animatedBirdData()
         provider.request(MultiTarget(Giphy.upload(gif: data)),
-                                  queue: DispatchQueue.main,
+                                  callbackQueue: DispatchQueue.main,
                                   progress: progressClosure,
                                   completion: progressCompletionClosure)
     }
     
     func downloadMoyaLogo() {
         provider.request(MultiTarget(GitHubUserContent.downloadMoyaWebContent("logo_github.png")),
-                                              queue: DispatchQueue.main,
+                                              callbackQueue: DispatchQueue.main,
                                               progress: progressClosure,
                                               completion: progressCompletionClosure)
     }