+++ Sources/Moya/MoyaProvider.swift
@@ -54,13 +54,13 @@ open class MoyaProvider<Target: TargetType> {
     open internal(set) var inflightRequests: [Endpoint<Target>: [Moya.Completion]] = [:]
 
     /// Propagated to Alamofire as callback queue. If nil - main queue will be used.
-    let queue: DispatchQueue?
+    let callbackQueue: DispatchQueue?
 
     /// Initializes a provider.
     public init(endpointClosure: @escaping EndpointClosure = MoyaProvider.defaultEndpointMapping,
                 requestClosure: @escaping RequestClosure = MoyaProvider.defaultRequestMapping,
                 stubClosure: @escaping StubClosure = MoyaProvider.neverStub,
-                queue: DispatchQueue? = nil,
+                callbackQueue: DispatchQueue? = nil,
                 manager: Manager = MoyaProvider<Target>.defaultAlamofireManager(),
                 plugins: [PluginType] = [],
                 trackInflights: Bool = false) {
@@ -71,7 +71,7 @@ open class MoyaProvider<Target: TargetType> {
         self.manager = manager
         self.plugins = plugins
         self.trackInflights = trackInflights
-        self.queue = queue
+        self.callbackQueue = callbackQueue
     }
 
     /// Returns an `Endpoint` based on the token, method, and parameters by invoking the `endpointClosure`.
@@ -82,14 +82,14 @@ open class MoyaProvider<Target: TargetType> {
     /// Designated request-making method. Returns a `Cancellable` token to cancel the request later.
     @discardableResult
     open func request(_ target: Target, completion: @escaping Moya.Completion) -> Cancellable {
-        return self.request(target, queue: queue, completion: completion)
+        return self.request(target, callbackQueue: callbackQueue, completion: completion)
     }
 
     /// Designated request-making method with queue option. Returns a `Cancellable` token to cancel the request later.
     @discardableResult
-    open func request(_ target: Target, queue: DispatchQueue?, progress: Moya.ProgressBlock? = nil, completion: @escaping Moya.Completion) -> Cancellable {
-        let callbackQueue = queue ?? self.queue
-        return requestNormal(target, queue: callbackQueue, progress: progress, completion: completion)
+    open func request(_ target: Target, callbackQueue: DispatchQueue?, progress: Moya.ProgressBlock? = nil, completion: @escaping Moya.Completion) -> Cancellable {
+        let callbackQueue = callbackQueue ?? self.callbackQueue
+        return requestNormal(target, callbackQueue: callbackQueue, progress: progress, completion: completion)
     }
 
     /// When overriding this method, take care to `notifyPluginsOfImpendingStub` and to perform the stub using the `createStubFunction` method.