+++ Tests/MoyaProviderIntegrationTests.swift
@@ -95,11 +95,11 @@ class MoyaProviderIntegrationTests: QuickSpec {
 
                     it("uses other background queue") {
                         var isMainThread: Bool?
-                        let queue = DispatchQueue(label: "background_queue", attributes: .concurrent)
+                        let callbackQueue = DispatchQueue(label: "background_queue", attributes: .concurrent)
                         let target: GitHub = .zen
 
                         waitUntil { done in
-                            provider.request(target, queue:queue) { _ in
+                            provider.request(target, callbackQueue: callbackQueue) { _ in
                                 isMainThread = Thread.isMainThread
                                 done()
                             }